Move types and methods out of types.go
